requirements.txt
fastapi==0.95.0
uvicorn==0.28.0
numpy>=1.20.3,<1.24
tensorflow==2.13.0
sentinelhub
rasterio
matplotlib
scikit-learn
python-dotenv==1.0.1
requests==2.31.0
fake-rpi
pydantic==1.10.7
pandas
pytest
pytest-mock
pytest-cov
PyYAML
fastapi-cache2
aioredis
lark
httpx
redis==5.0.1
passlib==1.7.4
email-validator==2.1.0
numba>=0.56.4,<0.57
imblearn
seaborn